pyproject.toml
[project]
name = "matias-live-cv"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi[standard] (>=0.129.0,<0.130.0)",
    "sqlalchemy[asyncio] (>=2.0.46,<3.0.0)",
    "pwdlib[argon2] (>=0.3.0,<0.4.0)",
    "pyjwt (>=2.11.0,<3.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "sqladmin (>=0.23.0,<0.24.0)",
    "itsdangerous (>=2.2.0,<3.0.0)",
    "pgvector (<0.4)",
    "langchain (>=1.2.10,<2.0.0)",
    "langchain-openai (>=1.1.10,<2.0.0)",
    "langchain-groq (>=1.1.2,<2.0.0)",
    "langchain-postgres (>=0.0.17,<0.0.18)",
    "pypdf (>=6.7.2,<7.0.0)",
    "langchain-community (>=0.4.1,<0.5.0)",
    "langchain-text-splitters (>=1.1.1,<2.0.0)",
    "slowapi (>=0.1.9,<0.2.0)",
    "boto3 (>=1.42.54,<2.0.0)",
    "pillow (>=12.1.1,<13.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false